fixup! Decompose channel: extract transport and static modules
