+++ lib/channel.js
@@ -3,7 +3,8 @@
 const http = require('http');
 const metautil = require('metautil');
 const transport = require('./transport.js');
-const { MIME_TYPES, HEADERS, COOKIE_DELETE, COOKIE_HOST, TOKEN } = transport;
+const { MIME_TYPES, HEADERS } = transport.http;
+const { COOKIE_DELETE, COOKIE_HOST, TOKEN } = transport.http;
 
 class Session {
   constructor(token, channel, data) {
