+++ lib/transport.js
@@ -1,29 +1,5 @@
 'use strict';
 
-const MIME_TYPES = {
-  html: 'text/html; charset=UTF-8',
-  json: 'application/json; charset=UTF-8',
-  js: 'application/javascript; charset=UTF-8',
-  css: 'text/css',
-  png: 'image/png',
-  ico: 'image/x-icon',
-  svg: 'image/svg+xml',
+module.exports = {
+  http: require('./http.js'),
 };
-
-const HEADERS = {
-  'X-XSS-Protection': '1; mode=block',
-  'X-Content-Type-Options': 'nosniff',
-  'Strict-Transport-Security': 'max-age=31536000; includeSubdomains; preload',
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'POST, GET, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type',
-};
-
-const TOKEN = 'token';
-const EPOCH = 'Thu, 01 Jan 1970 00:00:00 GMT';
-const FUTURE = 'Fri, 01 Jan 2100 00:00:00 GMT';
-const LOCATION = 'Path=/; Domain';
-const COOKIE_DELETE = `${TOKEN}=deleted; Expires=${EPOCH}; ${LOCATION}=`;
-const COOKIE_HOST = `Expires=${FUTURE}; ${LOCATION}`;
-
-module.exports = { MIME_TYPES, HEADERS, COOKIE_DELETE, COOKIE_HOST, TOKEN };
